project refactoring
